Rename validate and convert modules
There are endpoints and functions named "validate" and "convert" so this
prevents clashes.
