+++ test/unit/test_validation.py
@@ -5,8 +5,7 @@ from pathlib import Path
 from freezegun import freeze_time
 import pytest
 
-from app import ags
-from app import validate
+from app import validation
 from test.fixtures import (DICTIONARIES, FROZEN_TIME, ISVALID_RSP_DATA)
 from test.fixtures_json import JSON_RESPONSES
 from test.fixtures_plain_text import PLAIN_TEXT_RESPONSES
@@ -40,7 +39,7 @@ def test_validate_default_checker():
         'valid': True}
 
     # Act
-    response = validate.validate(filename, checkers=[mock_check_ags])
+    response = validation.validate(filename, checkers=[mock_check_ags])
 
     # Assert
     response == expected
@@ -62,7 +61,7 @@ def test_validate_bgs_checker():
         'valid': False}
 
     # Act
-    response = validate.validate(filename, checkers=[mock_check_bgs])
+    response = validation.validate(filename, checkers=[mock_check_bgs])
 
     # Assert
     response == expected
@@ -84,7 +83,7 @@ def test_validate_both_checkers():
         'valid': False}
 
     # Act
-    response = validate.validate(filename, checkers=[mock_check_bgs, mock_check_ags])
+    response = validation.validate(filename, checkers=[mock_check_bgs, mock_check_ags])
 
     # Assert
     response == expected
@@ -106,7 +105,7 @@ def test_validate_non_ags():
         'valid': False}
 
     # Act
-    response = validate.validate(filename)
+    response = validation.validate(filename)
 
     # Assert
     response == expected
@@ -120,7 +119,7 @@ def test_validate(filename, expected):
     filename = TEST_FILE_DIR / filename
 
     # Act
-    response = validate.validate(filename)
+    response = validation.validate(filename)
 
     # Assert
     # Check that metadata fields are correct
@@ -135,8 +134,8 @@ def test_validate_custom_dictionary(dictionary):
     filename = TEST_FILE_DIR / 'example_ags.ags'
 
     # Act
-    response = validate.validate(filename,
-                                 standard_AGS4_dictionary=dictionary)
+    response = validation.validate(filename,
+                                   standard_AGS4_dictionary=dictionary)
 
     # Assert
     assert response['filename'] == 'example_ags.ags'
@@ -150,12 +149,12 @@ def test_validate_custom_dictionary_bad_file():
 
     # Act
     with pytest.raises(ValueError) as err:
-        validate.validate(filename, standard_AGS4_dictionary=dictionary)
+        validation.validate(filename, standard_AGS4_dictionary=dictionary)
 
     # Assert
     message = str(err.value)
     assert 'dictionary' in message
-    for key in ags.STANDARD_DICTIONARIES:
+    for key in validation.STANDARD_DICTIONARIES:
         assert key in message
 
 
@@ -165,7 +164,7 @@ def test_is_valid(filename, expected):
     filename = TEST_FILE_DIR / filename
 
     # Act
-    result = validate.is_valid(filename)
+    result = validation.is_valid(filename)
 
     # Assert
     assert result == expected
@@ -177,8 +176,8 @@ def test_is_valid_custom_dictionary(dictionary):
     filename = TEST_FILE_DIR / 'example_ags.ags'
 
     # Act
-    result = validate.is_valid(filename,
-                               standard_AGS4_dictionary=dictionary)
+    result = validation.is_valid(filename,
+                                 standard_AGS4_dictionary=dictionary)
 
     # Assert
     assert result
@@ -193,7 +192,7 @@ def test_to_plain_text(filename):
     expected = PLAIN_TEXT_RESPONSES[filename]
 
     # Act
-    text = validate.to_plain_text(response)
+    text = validation.to_plain_text(response)
 
     # Assert
     assert text.strip() == expected.strip()
